+++ packages/patrol/lib/src/platform/contracts/mobile_automator_client.dart
@@ -63,6 +63,10 @@ class MobileAutomatorClient {
     return _sendRequest('openUrl', request.toJson());
   }
 
+  Future<void> sendKeyboardEnter() {
+    return _sendRequest('sendKeyboardEnter');
+  }
+
   Future<void> pressVolumeUp() {
     return _sendRequest('pressVolumeUp');
   }